// backend/src/modules/auth/auth.schemas.ts
import { z } from 'zod';

import type { LoginRequest, RegisterRequest } from '../../autogenerated/types/index.js';

export const registerRequestSchema: z.ZodType<RegisterRequest> = z.object({
  email: z.string().email(),
  name: z.string().min(1),
  password: z.string().min(8),
});

export const loginRequestSchema: z.ZodType<LoginRequest> = z.object({
  email: z.string().email(),
  password: z.string().min(1),
});

export type RegisterRequestInput = z.infer<typeof registerRequestSchema>;
export type LoginRequestInput = z.infer<typeof loginRequestSchema>;